NOTE: we need to tag stuff with @RemoteServiceRelativePath(...) because GWT hates us

+++ src/punishers/thirst/client/WaterFountainService.java
@@ -6,11 +6,11 @@ import com.google.gwt.user.client.rpc.RemoteService;
 
 public interface WaterFountainService extends RemoteService {
 	
-	public void addWaterFountainToFavs(int id) throws NotLoggedInException;
+	public void addWaterFountainToFavs(long id) throws NotLoggedInException;
 	
-	public void removeWaterFountainFromFavs(int id) throws NotLoggedInException;
+	public void removeWaterFountainFromFavs(long id) throws NotLoggedInException;
 	
 	public String[] getFavWaterFountains() throws NotLoggedInException;
 	
-	public Set<Integer> getAllIds() throws NotLoggedInException;
+	public Set<Long> getAllIds() throws NotLoggedInException;
 }
