+++ src/punishers/thirst/client/WaterFountainServiceAsync.java
@@ -6,12 +6,12 @@ import com.google.gwt.user.client.rpc.AsyncCallback;
 
 public interface WaterFountainServiceAsync {
 	
-	public void addWaterFountainToFavs(int id, AsyncCallback<Void> async);
+	public void addWaterFountainToFavs(long id, AsyncCallback<Void> async);
 	
-	public void removeWaterFountainFromFavs(int id, AsyncCallback<Void> async);
+	public void removeWaterFountainFromFavs(long id, AsyncCallback<Void> async);
 	
 	public void getFavWaterFountains(AsyncCallback<String[]> async);
 
-	public void getAllIds(AsyncCallback<Set<Integer>> callback);
+	public void getAllIds(AsyncCallback<Set<Long>> callback);
 
 }
